requirements.txt
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
yfinance>=0.2
pytest>=8.0
matplotlib>=3.8
